refactor: 🛠️ implicitify blog-post libs
