--- libs/blog-post-feature-detail/README.md
@@ -1,7 +1,0 @@
-# blog-post-feature-detail
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test blog-post-feature-detail` to execute the unit tests.
