--- libs/blog-post-feature-detail/project.json
@@ -1,20 +1,0 @@
-{
-  "name": "blog-post-feature-detail",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "sourceRoot": "libs/blog-post-feature-detail/src",
-  "prefix": "mc",
-  "tags": ["scope:blog-post", "type:feature"],
-  "targets": {
-    "lint": {
-      "executor": "@nx/eslint:lint"
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/libs/blog-post-feature-detail"],
-      "options": {
-        "jestConfig": "libs/blog-post-feature-detail/jest.config.ts"
-      }
-    }
-  }
-}
