--- libs/blog-post-ui/cypress/support/component.ts
@@ -1,1 +1,0 @@
-import '@marmicode/testing/cypress-ct-commands';
