--- libs/blog-post-ui/cypress/support/styles.scss
@@ -1,6 +1,0 @@
-@import '@fortawesome/fontawesome-svg-core/styles.css';
-@import '../../../../apps/marmicode/src/styles.scss';
-
-body {
-    background-color: white !important;
-}
