--- libs/blog-post-ui/jest.config.ts
@@ -1,24 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'blog-post-ui',
-  preset: '../../jest.preset.js',
-  setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
-  globals: {},
-  coverageDirectory: '../../coverage/libs/blog-post-ui',
-  snapshotSerializers: [
-    'jest-preset-angular/build/serializers/no-ng-attributes',
-    'jest-preset-angular/build/serializers/ng-snapshot',
-    'jest-preset-angular/build/serializers/html-comment',
-  ],
-  transform: {
-    '^.+.(ts|mjs|js|html)$': [
-      'jest-preset-angular',
-      {
-        stringifyContentPathRegex: '\\.(html|svg)$',
-
-        tsconfig: '<rootDir>/tsconfig.spec.json',
-      },
-    ],
-  },
-  transformIgnorePatterns: ['node_modules/(?!.*.mjs$)'],
-};
