--- libs/blog-post-ui/src/lib/social/share-buttons.component.scss
@@ -1,1 +1,0 @@
-@use 'ngx-sharebuttons/themes/outline';
