--- libs/blog-post-ui/src/test-setup.ts
@@ -1,2 +1,0 @@
-import '@marmicode/shared-utils/testing/test-setup';
-import 'jest-preset-angular/setup-jest';
