--- libs/blog-post-ui/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "files": [],
-  "include": [],
-  "references": [
-    {
-      "path": "./tsconfig.lib.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
-    },
-    {
-      "path": "./.storybook/tsconfig.json"
-    },
-    {
-      "path": "./cypress/tsconfig.json"
-    }
-  ],
-  "compilerOptions": {
-    "target": "es2020"
-  }
-}
