+++ libs/blog-post/feature-detail/lib/blog-post-detail-page.component.ts
