+++ libs/blog-post/feature-detail/lib/blog-post-to-meta.spec.ts
