+++ libs/blog-post/ui/.storybook/preview.js
@@ -4,4 +4,4 @@ import { withKnobs } from '@storybook/addon-knobs';
 addDecorator(withKnobs);
 
 // eslint-disable-next-line @nx/enforce-module-boundaries
-import '../../../.storybook/preview';
+import '../../../../.storybook/preview';
