+++ libs/blog-post/ui/.storybook/tsconfig.json
@@ -3,7 +3,9 @@
   "compilerOptions": {
     "emitDecoratorMetadata": true
   },
-  "files": ["../src/polyfills.ts"],
+  "files": [
+    "../src/polyfills.ts"
+  ],
   "exclude": ["../**/*.spec.ts", "jest.config.ts"],
   "include": [
     "../src/**/*.stories.ts",
