+++ libs/blog-post/ui/cypress/support/component.ts
@@ -1,0 +1,1 @@
+import 'libs/testing/src/cypress-ct-commands';
