+++ libs/blog-post/ui/cypress/support/styles.scss
@@ -1,0 +1,6 @@
+@import 'styles.css';
+@import '../../../../../apps/marmicode/src/styles';
+
+body {
+    background-color: white !important;
+}
