+++ libs/blog-post/ui/project.json
@@ -1,8 +1,8 @@
 {
   "name": "blog-post-ui",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
-  "sourceRoot": "libs/blog-post-ui/src",
+  "sourceRoot": "{projectRoot}/src",
   "prefix": "mc",
   "tags": ["scope:blog-post", "type:ui"],
   "targets": {
@@ -11,16 +11,16 @@
     },
     "test": {
       "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/libs/blog-post-ui"],
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/blog-post-ui/jest.config.ts"
+        "jestConfig": "{projectRoot}/jest.config.ts"
       }
     },
     "storybook": {
       "executor": "@storybook/angular:start-storybook",
       "options": {
         "port": 4400,
-        "configDir": "libs/blog-post-ui/.storybook",
+        "configDir": "{projectRoot}/.storybook",
         "browserTarget": "blog-post-ui:build-storybook",
         "compodoc": false
       },
@@ -33,8 +33,8 @@
     "build-storybook": {
       "executor": "@storybook/angular:build-storybook",
       "options": {
-        "outputDir": "dist/storybook/blog-post-ui",
-        "configDir": "libs/blog-post-ui/.storybook",
+        "outputDir": "dist/storybook/{projectRoot}",
+        "configDir": "{projectRoot}/.storybook",
         "browserTarget": "blog-post-ui:build-storybook",
         "compodoc": false
       },
@@ -49,9 +49,9 @@
       "options": {
         "outputPath": "does-not-exist",
         "index": "does-not-exist",
-        "polyfills": ["libs/blog-post-ui/src/polyfills.ts"],
-        "tsConfig": "libs/blog-post-ui/tsconfig.lib.json",
-        "styles": ["libs/blog-post-ui/cypress/support/styles.scss"],
+        "polyfills": ["{projectRoot}/src/polyfills.ts"],
+        "tsConfig": "{projectRoot}/tsconfig.lib.json",
+        "styles": ["{projectRoot}/cypress/support/styles.scss"],
         "stylePreprocessorOptions": {
           "includePaths": [
             "node_modules"
@@ -64,7 +64,7 @@
     "component-test": {
       "executor": "@nx/cypress:cypress",
       "options": {
-        "cypressConfig": "libs/blog-post-ui/cypress.config.js",
+        "cypressConfig": "libs/blog-post/ui/cypress.config.js",
         "testingType": "component",
         "skipServe": true,
         "devServerTarget": "blog-post-ui:component-test-build",
@@ -72,7 +72,7 @@
       }
     },
     "upload-percy-snapshots": {
-      "command": "tools/upload-percy-snapshots.sh libs/blog-post-ui"
+      "command": "tools/upload-percy-snapshots.sh {projectRoot}"
     }
   }
 }
