+++ libs/blog-post/ui/src/lib/social/share-buttons.component.scss
@@ -1,0 +1,1 @@
+@use '../../../../../../node_modules/ngx-sharebuttons/themes/outline';
