+++ libs/blog-post/ui/tsconfig.spec.json
@@ -6,6 +6,8 @@
     "types": ["jest", "node"],
     "target": "es2016"
   },
-  "files": ["src/test-setup.ts"],
-  "include": ["**/*.spec.ts", "**/*.test.ts", "**/*.d.ts", "jest.config.ts"]
+  "files": [
+    "src/test-setup.ts"
+  ],
+  "include": ["**/*.spec.ts", "**/*.test.ts", "**/*.d.ts"]
 }
