+++ tsconfig.base.json
@@ -23,10 +23,10 @@
       "@marmicode/block/api": ["libs/block/api/index.ts"],
       "@marmicode/block/core": ["libs/block/core/index.ts"],
       "@marmicode/block/ui": ["libs/block/ui/src/index.ts"],
-      "@marmicode/blog-post-feature-detail": [
-        "libs/blog-post-feature-detail/src/index.ts"
+      "@marmicode/blog-post/feature-detail": [
+        "libs/blog-post/feature-detail/index.ts"
       ],
-      "@marmicode/blog-post-ui": ["libs/blog-post-ui/src/index.ts"],
+      "@marmicode/blog-post/ui": ["libs/blog-post/ui/index.ts"],
       "@marmicode/contentful-infra": ["libs/contentful-infra/src/index.ts"],
       "@marmicode/learning-map-feature-tree": [
         "libs/learning-map-feature-tree/src/index.ts"
